card-detector/requirements.txt
streamlit
azure-core
python-dotenv